# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dhan Algo Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get all orders
      description: Fetch all orders from Dhan
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place a new order
      description: Place a new order on Dhan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderBody"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
...
  /trades/history:
    get:
      operationId: getTradeHistory
      tags: [trades]
      summary: Get trade history
      description: Fetch historical trades within a date range
      parameters:
        - name: fromDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
...
  /strategies:
    get:
      operationId: getStrategies
      tags: [strategies]
      summary: Get all strategies
      description: List all saved trading strategies
      responses:
        "200":
          description: Strategy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
    post:
      operationId: createStrategy
      tags: [strategies]
      summary: Create a strategy
      description: Create a new trading strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStrategyBody"
      responses:
        "201":
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}:
    get:
      operationId: getStrategy
      tags: [strategies]
      summary: Get a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
    patch:
      operationId: updateStrategy
      tags: [strategies]
      summary: Update a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStrategyBody"
      responses:
        "200":
          description: Strategy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
    delete:
      operationId: deleteStrategy
      tags: [strategies]
      summary: Delete a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Strategy deleted
  /strategies/{id}/toggle:
    post:
      operationId: toggleStrategy
      tags: [strategies]
      summary: Toggle strategy active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}/execute:
    post:
      operationId: executeStrategy
      tags: [strategies]
      summary: Execute a strategy manually
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyExecutionResult"
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        orderId:
          type: string
        correlationId:
          type: string
        orderStatus:
          type: string
        transactionType:
          type: string
        exchangeSegment:
          type: string
        productType:
          type: string
        orderType:
          type: string
        validity:
          type: string
        tradingSymbol:
          type: string
        securityId:
          type: string
        quantity:
          type: integer
        disclosedQuantity:
          type: integer
        price:
          type: number
        triggerPrice:
          type: number
        afterMarketOrder:
          type: boolean
        boProfitValue:
          type: number
        boStopLossValue:
          type: number
        legName:
          type: string
        createTime:
          type: string
        updateTime:
          type: string
        exchangeTime:
          type: string
        drvExpiryDate:
          type: string
        drvOptionType:
          type: string
        drvStrikePrice:
          type: number
        omsErrorCode:
          type: string
        omsErrorDescription:
          type: string
        filledQty:
          type: integer
        averagePrice:
          type: number
      required:
        - orderId
        - orderStatus
        - transactionType
        - exchangeSegment
        - productType
        - orderType
        - securityId
        - quantity
        - price
    PlaceOrderBody:
      type: object
      properties:
        securityId:
          type: string
        exchangeSegment:
          type: string
        transactionType:
          type: string
          enum: [BUY, SELL]
        quantity:
          type: integer
        orderType:
          type: string
          enum: [MARKET, LIMIT, SL, SLM]
        productType:
          type: string
          enum: [INTRA, CNC, MARGIN, CO, BO]
        price:
          type: number
        triggerPrice:
          type: number
        disclosedQuantity:
          type: integer
        afterMarketOrder:
          type: boolean
        validity:
          type: string
          enum: [DAY, IOC]
        boProfitValue:
          type: number
        boStopLossValue:
          type: number
        tag:
          type: string
      required:
        - securityId
        - exchangeSegment
        - transactionType
        - quantity
        - orderType
        - productType
        - price
...
    OrderResponse:
      type: object
      properties:
        orderId:
          type: string
        status:
          type: string
        message:
          type: string
      required:
        - orderId
        - status
...
    Trade:
      type: object
      properties:
        orderId:
          type: string
        exchangeOrderId:
          type: string
        exchangeTradeId:
          type: string
        transactionType:
          type: string
        exchangeSegment:
          type: string
        productType:
          type: string
        orderType:
          type: string
        tradingSymbol:
          type: string
        securityId:
          type: string
        quantity:
          type: integer
        tradedQuantity:
          type: integer
        tradedPrice:
          type: number
        createTime:
          type: string
        updateTime:
          type: string
        drvExpiryDate:
          type: string
        drvOptionType:
          type: string
        drvStrikePrice:
          type: number
      required:
        - orderId
        - transactionType
        - tradingSymbol
        - securityId
        - tradedQuantity
        - tradedPrice
...
    Strategy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [scalping, swing, positional, options, custom]
        status:
          type: string
          enum: [active, paused, stopped]
        securityId:
          type: string
        tradingSymbol:
          type: string
        exchangeSegment:
          type: string
        transactionType:
          type: string
        productType:
          type: string
        orderType:
          type: string
        quantity:
          type: integer
        entryPrice:
          type: number
        stopLoss:
          type: number
        target:
          type: number
        trailingStopLoss:
          type: number
        maxPositions:
          type: integer
        maxLossPerDay:
          type: number
        maxProfitPerDay:
          type: number
        timeframeMinutes:
          type: integer
        entryConditions:
          type: string
        exitConditions:
          type: string
        totalTrades:
          type: integer
        winTrades:
          type: integer
        lossTrades:
          type: integer
        totalPnl:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - status
        - securityId
        - tradingSymbol
        - exchangeSegment
        - transactionType
        - productType
        - orderType
        - quantity
        - createdAt
        - updatedAt
    CreateStrategyBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [scalping, swing, positional, options, custom]
        securityId:
          type: string
        tradingSymbol:
          type: string
        exchangeSegment:
          type: string
        transactionType:
          type: string
        productType:
          type: string
        orderType:
          type: string
        quantity:
          type: integer
        entryPrice:
          type: number
        stopLoss:
          type: number
        target:
          type: number
        trailingStopLoss:
          type: number
        maxPositions:
          type: integer
        maxLossPerDay:
          type: number
        maxProfitPerDay:
          type: number
        timeframeMinutes:
          type: integer
        entryConditions:
          type: string
        exitConditions:
          type: string
      required:
        - name
        - type
        - securityId
        - tradingSymbol
        - exchangeSegment
        - transactionType
        - productType
        - orderType
        - quantity
    UpdateStrategyBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [scalping, swing, positional, options, custom]
        status:
          type: string
          enum: [active, paused, stopped]
        securityId:
          type: string
        tradingSymbol:
          type: string
        exchangeSegment:
          type: string
        transactionType:
          type: string
        productType:
          type: string
        orderType:
          type: string
        quantity:
          type: integer
        entryPrice:
          type: number
        stopLoss:
          type: number
        target:
          type: number
        trailingStopLoss:
          type: number
        maxPositions:
          type: integer
        maxLossPerDay:
          type: number
        maxProfitPerDay:
          type: number
        timeframeMinutes:
          type: integer
        entryConditions:
          type: string
        exitConditions:
          type: string
    StrategyExecutionResult:
      type: object
      properties:
        strategyId:
          type: integer
        orderId:
          type: string
        status:
          type: string
        message:
          type: string
        executedAt:
          type: string
          format: date-time
      required:
        - strategyId
        - status
        - message
        - executedAt
...