# lib/api-zod/src/generated/types/createStrategyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dhan Algo Trading Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateStrategyBodyType } from "./createStrategyBodyType";

export interface CreateStrategyBody {
  name: string;
  description?: string;
  type: CreateStrategyBodyType;
  securityId: string;
  tradingSymbol: string;
  exchangeSegment: string;
  transactionType: string;
  productType: string;
  orderType: string;
  quantity: number;
  entryPrice?: number;
  stopLoss?: number;
  target?: number;
  trailingStopLoss?: number;
  maxPositions?: number;
  maxLossPerDay?: number;
  maxProfitPerDay?: number;
  timeframeMinutes?: number;
  entryConditions?: string;
  exitConditions?: string;
}